requirements.txt
psycopg2-binary>=2.9.9,<3.0
PyYAML>=6.0.1,<7.0
requests>=2.31.0,<3.0
pytest>=8.0.0,<9.0
pandas>=2.2.0,<3.0
numpy>=1.26.0,<3.0
scikit-learn>=1.4.0,<2.0
joblib>=1.3.0,<2.0
